# docs/openapi/repository-api-v1.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "description": "Error detail object for the public API error envelope.",
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured metadata to help clients debug the error.",
            "title": "Details"
          },
          "message": {
            "description": "Human-readable summary of the failure.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "Standardized error envelope shared across HTTP endpoints.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody",
            "description": "Normalized error payload."
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "RelationshipSelectorResponse": {
        "additionalProperties": false,
        "description": "Authored relationship selector preserved exactly as published.",
        "properties": {
          "markers": {
            "items": {
              "type": "string"
            },
            "title": "Markers",
            "type": "array"
          },
          "optional": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optional"
          },
          "slug": {
            "description": "Stable public slug of the related skill.",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
            "title": "Slug",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Exact related version when the selector targets one immutable version.",
            "title": "Version"
          },
          "version_constraint": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Authored dependency constraint when the selector is version-ranged.",
            "title": "Version Constraint"
          }
        },
        "required": [
          "slug"
        ],
        "title": "RelationshipSelectorResponse",
        "type": "object"
      },
...
      "SkillRelationshipBatchItemResponse": {
        "description": "One ordered direct relationship lookup result.",
        "properties": {
          "coordinate": {
            "$ref": "#/components/schemas/SkillVersionCoordinateRequest"
          },
          "relationships": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SkillRelationshipEdgeResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationships"
          },
          "status": {
            "enum": [
              "found",
              "not_found"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "coordinate"
        ],
        "title": "SkillRelationshipBatchItemResponse",
        "type": "object"
      },
      "SkillRelationshipBatchRequest": {
        "additionalProperties": false,
        "description": "Ordered direct relationship query over immutable source versions.",
        "properties": {
          "coordinates": {
            "items": {
              "$ref": "#/components/schemas/SkillVersionCoordinateRequest"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Coordinates",
            "type": "array"
          },
          "edge_types": {
            "items": {
              "enum": [
                "depends_on",
                "extends",
                "conflicts_with",
                "overlaps_with"
              ],
              "type": "string"
            },
            "title": "Edge Types",
            "type": "array"
          }
        },
        "required": [
          "coordinates"
        ],
        "title": "SkillRelationshipBatchRequest",
        "type": "object"
      },
      "SkillRelationshipBatchResponse": {
        "description": "Ordered batch response for direct immutable relationship reads.",
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SkillRelationshipBatchItemResponse"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "SkillRelationshipBatchResponse",
        "type": "object"
      },
      "SkillRelationshipEdgeResponse": {
        "description": "One direct authored relationship edge from an immutable source version.",
        "properties": {
          "edge_type": {
            "enum": [
              "depends_on",
              "extends",
              "conflicts_with",
              "overlaps_with"
            ],
            "title": "Edge Type",
            "type": "string"
          },
          "selector": {
            "$ref": "#/components/schemas/RelationshipSelectorResponse"
          },
          "target_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SkillVersionReferenceResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "edge_type",
          "selector"
        ],
        "title": "SkillRelationshipEdgeResponse",
        "type": "object"
      },
...
      "SkillVersionCoordinateRequest": {
        "additionalProperties": false,
        "description": "Exact immutable slug/version coordinate.",
        "properties": {
          "slug": {
            "description": "Stable public slug of the requested skill.",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
            "title": "Slug",
            "type": "string"
          },
          "version": {
            "description": "Exact immutable semantic version of the requested skill.",
            "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "version"
        ],
        "title": "SkillVersionCoordinateRequest",
        "type": "object"
      },
...
      "SkillVersionReferenceResponse": {
        "description": "Compact exact version reference.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "lifecycle_status": {
            "enum": [
              "published",
              "deprecated",
              "archived"
            ],
            "title": "Lifecycle Status",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "trust_tier": {
            "enum": [
              "untrusted",
              "internal",
              "verified"
            ],
            "title": "Trust Tier",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "version",
          "name",
          "description",
          "tags",
          "lifecycle_status",
          "trust_tier",
          "published_at"
        ],
        "title": "SkillVersionReferenceResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Registry-first API for immutable skill publication, indexed search candidate\nretrieval, direct relationship reads, exact version retrieval, artifact streaming,\nand version listing.\n\nThe server owns data-local registry operations only. Prompt interpretation,\nreranking, dependency solving, final selection, lock generation, and execution\nplanning remain client-owned behavior outside this API boundary.",
    "title": "aptitude-server Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/resolution/relationships:batch": {
      "post": {
        "description": "Return authored direct `depends_on` and `extends` relationships for exact immutable source versions. This route does not expand transitive graphs, select versions for constraints, or emit solved dependency closures.",
        "operationId": "batchGetDirectSkillRelationships",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillRelationshipBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "results": [
                    {
                      "coordinate": {
                        "slug": "python.lint",
                        "version": "1.2.3"
                      },
                      "relationships": [
                        {
                          "edge_type": "depends_on",
                          "selector": {
                            "markers": [
                              "linux",
                              "gpu"
                            ],
                            "optional": true,
                            "slug": "python.base",
                            "version_constraint": ">=1.0.0,<2.0.0"
                          }
                        },
                        {
                          "edge_type": "extends",
                          "selector": {
                            "markers": [],
                            "slug": "python.base",
                            "version": "1.0.0"
                          },
                          "target_version": {
                            "description": "Base runtime skill",
                            "lifecycle_status": "published",
                            "name": "Python Base",
                            "published_at": "2026-03-01T08:30:00Z",
                            "slug": "python.base",
                            "tags": [
                              "python",
                              "runtime"
                            ],
                            "trust_tier": "verified",
                            "version": "1.0.0"
                          }
                        }
                      ],
                      "status": "found"
                    },
                    {
                      "coordinate": {
                        "slug": "python.missing",
                        "version": "9.9.9"
                      },
                      "status": "not_found"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/SkillRelationshipBatchResponse"
                }
              }
            },
            "description": "Ordered direct relationship results returned successfully."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "details": {
                      "errors": [
                        {
                          "input": "latest",
                          "loc": [
                            "path",
                            "version"
                          ],
                          "msg": "String should match pattern",
                          "type": "string_pattern_mismatch"
                        }
                      ]
                    },
                    "message": "Request validation failed."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The relationship query request is invalid."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read direct immutable relationships",
        "tags": [
          "resolution"
        ]
      }
    },
...
  }
}